TWINS-248: Update CommonSpecification, RelatedObjectsDTOv1, ResponseRelatedObjectsDTOv1, TierDTOv1, TierSearchDTOMapper, TierSearchService, TierSpecification, TwinclassSchemaDTOMapper

+++ core/src/main/java/org/twins/core/dto/rest/domain/TierDTOv1.java
@@ -35,7 +35,7 @@ public class TierDTOv1 {
     public Boolean custom;
 
     @Schema(description = "attachments storage quota count")
-    public Long attachmentsStorageQuotaCount;
+    public Integer attachmentsStorageQuotaCount;
 
     @Schema(description = "attachments storage quota size")
     public Long attachmentsStorageQuotaSize;
